+++ R/resmush_dir.R
@@ -28,7 +28,7 @@
 #' returned as well.
 #'
 #' @seealso
-#' [reSmush.it API](https://resmush.it/api) docs.
+#' [reSmush.it API](https://resmush.it/api/) docs.
 #'
 #' See [resmush_clean_dir()] to clean a directory of previous runs.
 #'